Убрана проверка ссылок в base_page
